Clean up some of the dir.props and dir.targets files.
Add a VerifyBuildTools target to dirs.proj to ensure check for build tools
before trying to build. I tried adding something similar to dir.targets but
there was some issues and it seemed to cause some perf hits checking in ever
project build. If we find that we run into issues we can reconsider adding that.

This change also updates build.proj to add a semaphore file to use to detect
if build tools have been installed or not.