Fixes UnmanagedMemoryStream.Read
When I ported UMS, I noticed a bug in code lines that copy memory. I fixed the bug in Write, but not in Read (because tests were not working properly). This commit fixes the bug in Read and fixes tests to be more robust.

Also, opportunistically, I changed some visibilities from internal to private.