Fix rebuild target when building from the root
Rebuild naturally calls Clean followed by Build but the _RestoreBuildTools
was set as the InitialTargets in the build.proj which means you restore tools
clean tools then try to build which doesn't work so well. This change set the
restore build tools target to before BuildAllProjects so that the target align
correctly.