Avoid scheduling to TaskScheduler.Current in dataflow tests
xunit uses a TaskScheduler that limits DOP, and it schedules the execution of tests to that scheduler.  As such, if a test blocks while waiting for another task that the test has scheduled back to TaskScheduler.Current, the test can deadlock.  This commit changes task scheduling to avoid going to TaskScheduler.Current, either by explicitly passing in a scheduler to StartNew/ContinueWith/etc. or by using Task.Run (which implicitly targets TaskScheduler.Default).