Add some messsages to our restore targets so they show in console output
There is always a plause when restoring build tools or package
dependencies so adding an output message so people can see what
is happening from the console where we show only minimal (aka High
importance messages).

This also adds a note to see the log file if there are test failures.