Updated the error handling for RunTestsForProject target
This will now give an error message that points people to the log
file to see the details of the failed tests as opposed to just the
command line of the test runner. The command line for the test
runner will still show up as a warning if people want to rerun it
independently.