Add vim swp files to the gitignore.
On a default windows LFN system and the *nix installs the swp file will be prefixed with a ., so the pattern takes that into account.  It can be relaxed if it's found to be under-matching.

This also only looks at the primary .swp, and doesn't encode the fallback logic (.swp => .swo => .swn => ... => .swa).  If needed, it can be added later.