Disable parallelization of Registry tests
As currently written, tests can conflict with each other in the state they access in the registry.  For example, tt appears that lots of tests use the "" subkey and then try to operate on different values using a interlocked counter to append a number to the test key name, but rather than using a shared static counter, each test class has its own static counter, so multiple tests from different classes running in parallel could get the same key name.

The registry tests in generall look like they could use some cleanup, both in general and specifically to avoid this kind of shared state.  But in the meantime, this commit disables the parallelization to avoid these random conflicts.