Fix VS build by adjusting relative path to packages folder
This was caused by PR #474, which moved the nuget packages
folder up one level. There were still two places conditioned on
$(BuildingInsideVisualStudio) that had the old path.

Note that the VS xunit runner isn't working at the moment, so
we don't even need this import at all, but keeping it in place
as a reference for how things last worked when running tests
on desktop. We may need to do it differently in the end. This
is just the minimal fix to unblock things.