Fix UnmanagedMemoryStream test failure
The test was assuming that new byte[] allocations would come later in
memory than the array allocation around which the stream was constructed.
When that assumption failed, so would the test.  This commit just rewrites
the test to be explicit about the values being tested relative to the
position of the fixed target array.