Replace dataflow "hardening" tests
The existing "hardening tests" were duplicative of a bunch of things covered once the rest of the dataflow tests were rewritten.  They also involved lots of timeouts that made them unstable, especially in a lab environment. I'm deleted the whole file and added some additional tests throughout the rest of the suite to make up for a few remaining deficiencies.  I also removed a bit of dead code from the product.  Code coverage is not at ~93% (after ignoring source like the copy of ConcurrentQueue that's included).