Pass -wait to xunit to pause after test run from VS
Allows tests to be run using Ctrl+F5 with a pause to
keep the window open so that the results can be seen.

This was previously blocked by missing -wait support
in the .NET Core xunit console runner, but that was
added with dotnet/buildtools@cc6a4b00b4fe4ceaf659c91.

Bump dependency version on xunit and buildtools to
pick it up.