Augment terminfo search with first-letter-as-hex
terminfo files are organized into subdirectories based on the first letter of the terminal name.  On Linux, the letter itself is used as the directory name.  On OS X, the hex value for the letter is used as the directory name.  This commit updates the logic that searches for the terminfo database file to try both locations.