Fix incremental build issue caused by _RestoreBuildTools target
While performing an incremental build via build.cmd/proj you can
get into a state where the _RestoreBuildTools target executes for
every project. While this doesn't break anything it does cause
as slowdown. The way to enter this state is to keep the packages
directory around for a while and get or make an update to dir.props
or .nuget\packages.config which are the inputs to the target. Then
you will execute the target everytime because nuget.exe nor
the built tools package will actually be updated as part of executing.
To fix this I've added a call to Touch at end of the target to ensure
that the last modified time is always greater on the output files
then the input files.