Cache evaluated console color strings
Each time a call to Console.ForegroundColor or Console.BackgroundColor is made on Unix, we access the terminfo database to get the right evaluation string, evaluate it, and then write out the result.  But there are only 16 possible colors for each of foreground and background, and these strings are idempotent.  This commit just adds a small cache to avoid recomputing these strings every time.