Implement System.IO.MemoryMappedFiles on Unix
This commit provides an initial implementation of the System.IO.MemoryMappedFiles assembly on Unix.

The main difference between the Windows and Unix implementations is that Windows has two primary APIs involved, CreateFileMapping and MapViewOfFile, and these two APIs are roughly approximated by the .NET static methods MemoryMappedFile.CreateFromFile/CreateNew/etc. and the instance methods MemoryMappedFile.CreateView*, respectively.  In contrast, Unix has these concepts combined into just the mmap function.  As such, on Unix I've implemented the static methods primarily to bundle up their arguments and store them into the "handle" that's returned; then when the view is created, mmap is called using the combined arguments from the creation of the MemoryMappedFile and from the creation of the MemoryMappedView.

The main gap in the implementation is that Windows has the notion of named maps (used primarily for IPC), a concept that doesn't have a direct equivalent on Unix.  Rather than try to approximate it, for now at least any attempts to use a map name result in a PlatformNotSupportedException.