Use MetadataStringDecoder from MetadataStringComparer
Remove the hand-rolled UTF-8 decoder that did not handle errors
correctly and defer to the MetadataStringDecoder in the same way that
MetadataReader.GetString already does. This allows the user-supplied
decoder to override the error handling (e.g. Exception fallback over
replacement fallback).

The challenge was that MetadataStringComparer is designed not to
allocate. The compromise implemented here is to greedily compare
as ascii and fall back to allocation and System.String comparison
when a non-ascii character is encountered.

It is assumed that both of the following are rare and there is no
allocation unless they both happen together:

1. Non-ascii strings in UTF-8 encoded portions of metadata
2. Comparing non-ascii System.String values to metadata.

If these assumptions don't hold over the long run, we can tweak this
further to decode N characters at a time to stack or pool-allocated
memory. That didn't seem to be worth the complexity witout evidence
of a real scenario and performance issue.

Fix #26