Rewrite disabled test to use portable API and re-enable it
The CanReadFromSameMemoryMappedPEReaderInParallel test was
using typeof(object).Assembly.Location (mscorlib) as a file
large enough to memory map. This API is not portable so
instead just concat a small test assembly onto itself in the
temp folder and open that.

Also removed the OuterLoop designation as the test is fast
and reliable enough to run always.

Fix #260