Use modern assemblies before classic assemblies in light-up
The metadata reader was using the desktop assembly identities for lightup,
which was causing us to miss types that have been brought to .NET Core
above and beyond our .NET 4.5 + Windows 8 portable profile.

We now light-up on FileStream and MemoryMappedFile API appropriately on
.NET Core as well.

This also eliminates the last of tests that were disabled during the push
to run all tests on .NET Core.

Fix #264
Fix #816