Add tests for System.Diagnostics.Debug
I could not find any existing tests in our internal conformance suite,
so I wrote some simple ones. The best way to do this for inner loop
testing seemed to be to add a hook to the debug class so that the
tests could change out the sink for debug events and do validation
that way.  This was easier now that we had specified a PAL.

Doing it this way means that we don't get to test much of the actual
code that writes to the debugger or pops up an Assert UI, but I don't
think we currently have a good strategy for testing things which want
to pop UI.