Replace Contract.Assert with Debug.Asserts/Fail
We previously agreed to unify on Debug.Assert rather than Contract.Assert, and @ellismg did a sweep through the code in the repo at the time to make the change.  Since then, some more Contract.Assert calls have snuck in.

This commit replaces another ~90 Contract.Assert usages, which is currently all of those left in the repo.