Fix a couple of DebuggerDisplayAttributes
AllocFreeConcurrentStack has a DebuggerDisplay attribute that a) uses a non-trivial expression (which is ill-advised) and b) uses field names that don't exist.  Further, as far as I can tell, there's no way in VS2015 to get the debugger to display a DebuggerDisplay for a static type like this; any attempts I've made yield an error "error CS0119: 'AllocFreeConcurrentStack<int>' is a type, which is not valid in the given context".  I've simply deleted the attribute.

KeyValuePairs in System.Collections.NonGeneric has included in its DDA the string 'Type = ""'.  I've just removed that portion of the DDA.