Fix RespectParentCancellation test to avoid reliance on timeout
The RespectParentCancellation test was using a two-second timeout.  This a) artifically delayed the test's execution, and b) made it susceptible to random failures if the loop being tested happened to complete within that time window.

This commit just rewrites that test to avoid relying on a timeout.  This removes the flakiness.  And since this particular test is executed three times with different inputs, it shaves ~6 seconds off the test suite's execution, which on my machine is a significant portion of the overall running time.