Fix UseWinRT heuristic to better handle app container roots
We were using FindFirstFileEx to get file attributes in order to determine if
the file/directory is accessible by Win32.  This is failing for root folders in
ApplicationData locations, even though they are accessible by Win32.  The
reason seems to be that FindFirstFileEx on a full path requires that the parent
have permission to enumerate contents, even though we've passed a full
path and have full access to that path (likely a bug in FindFirstFileEx).
Using GetFileAttributesEx is better in most cases since it is faster and does
not require permission to enumerate contents.  I've switched to using
GetFileAttributesEx first and only falling back to FindFirstFile when we might
have a reparse point that could be a placeholder file.