Give System.Linq.Tests.csproj a P2P reference to System.Linq.csproj
The System.Linq.Tests project was pulling in System.Linq.dll via a NuGet package, not via a P2P reference.  This meant that a) a potentially older version of System.Linq was being tested, and b) no PDB was being generated in the tests folder, resulting in no code coverage results.