Enable and fix System.IO.Pipes tests
The Pipes tests weren't running because the project was named ".BasicTests" instead of ".Tests".  This commit renames it so that it runs.

Further, while it wasn't running, the tests got broken.  The tests' packages.config wasn't pulling in some required packages, causing type load errors that were then masked as other failures.  This commit updates the packages.config to pull in the needed packages.

Once the tests were enabled, there was a large amount of Console.WriteLine spew.  This commit replaces all of the Console.WriteLines with asserts (or just removed them if they were superfluous).

Finally, as I was going through doing the investigation, I cleaned up a few random things.