Update Microsoft.Win32.Registry tests
Microsoft.Win32.Registry has a test which validates that Win32Exception class can handle
error codes with message strings greater than 256. This is done using one of the existing
error codes which has message length > 256. However this assumption is not guaranteed
to be true on all cultures making the test prone to failure.

Updating the test to first check that the assumption is indeed true. This might mean
we will not run this test at all on some cultures however that is ok as the goal is not to test this across cultures :).