Delete closed libraries sources and reference open sources.
This change removes the closed libraries sources for libraries which we are now building in the open.  Stub project files are left in the closed libraries which improt sources from the open for various build configurations.

Also includes these two changes which were not yet ported to open:
- [Serialization]: DCS/DCJS fails to serialize/deserialize XElement type - regression
- Change namespace of System.Xml.XPath.XmlDocument extensions

Validated by pre-checkin gauntlet build #150320.14337

[tfs-changeset: 1437136]