`ParallelQuery<float?>.Sum()` add up all elements.
The non-nullable float sum was able to add up elements, even when the
intermediate sum would not normally be representable,
by using a double during the actual summation, and casting on output.
The nullable float sum wasn't doing this, causing it to output different results
(even when the final result would be representable).

This commit also removes `checked` around floating-point addition,
 as it is ignored (returns Infinity instead).