Tweak an immutable collections test
Out of the 611 immutable collections tests, one was failing on Unix.  The test is verifying that an ImmutableSortedSet.Builder's enumerator returns values in the right sorted order.  The problem is that it's using a default comparer based on the current culture, and we've not yet implemented that support on Unix, instead defaulting all such comparers to use ordinal... this results in a different sort order, and the test fails.

Since the actual comparer used is immaterial to what the test is verifying, I've simple tweaked the test to use an ordinal comparer.  This then passes on both Windows and Unix.