Fix cross-plat issues in XmlResolver tests
The System.Xml.RW.XmlSystemPathResolver tests were using P/Invokes to Win32 for things that could have been done with existing .NET APIs.  This commit replaces those so that the tests can run on Unix.

The tests were also relying on an Empty.xml file which would need to be deployed with the tests.  Now that the tests are able to use System.IO.FileSystem, the commit removes that Empty.xml file and just generates a temporary empty file at run time.