Fix a bunch of "\r\n"s
...since it's no longer the only newline in town.

Note that there's still a bunch of \r\n usage, mostly in the XML tests, but in many cases it looks like it's intentional, testing all forms of newlines.  We can clean up any remaining stragglers on an as-needed basis... the ones fixed in this commit are those which were causing tests to fail or which seemd from quick inspection that they should be Environment.NewLine, such as in assert messages.