Use coveralls badge but link to original coverage report
Coveralls is a good service, but we'd like to see different breakdowns of code coverage (by assembly, namespace, etc.)  Coveralls is interested in this feature set but it's not currently available.  We like the overall badge % reporting though.  For the time being, use the badge but link to the original report data