Add env var to conditionally Debug.Write to stderr
Some tests use Debug.WriteLine to output diagnostic information.  On Windows this doesn't show up unless a debugger is attached.  But the current Debug.Write implementation on Unix includes outputting to stderr unconditionally, resulting in tests generating a lot of noisy output.  This commit adds a check for an environment variable; only if set will Debug.Write output to stderr.